refactor: update build script, adjust imports and typings, streamline testing structure
